# Log type evals perf when option is on