@heejaechang 2024-04-19
not sure whether `pyright` minify `js` files, if it does, log's function name such as `evaluateTypesForStatement` won't match original function name but will be minified name such as `Ke` or etc. but package source info such as `[Function '_load_pre_generated_assumption_rules' (sympy.core.assumptions)] (9305ms)` or `["dict([(('algebraic', True), se <shortened> " (..sympy.core.assumptions_generated) [42:21]] (9006ms)` should be shown correctly.